Split part of dockerd main to command under daemon
Allows using daemon/internal packages when creating dockerd
