+++ cmd/dockerd/builtins/genwinres_windows.go
@@ -1,0 +1,5 @@
+//go:generate go-winres make --arch=386,amd64,arm,arm64 --in=../../../cli/winresources/dockerd/winres.json --out=../../../cli/winresources/dockerd/resource
+
+package builtins
+
+import _ "github.com/docker/docker/cli/winresources/dockerd"
