+++ daemon/command/daemon_unix_test.go
@@ -1,6 +1,6 @@
-//go:build !windows
+//go:build unix
 
-package main
+package command
 
 import (
 	"testing"
