+++ daemon/command/docker.go
@@ -1,25 +1,21 @@
-package main
+package command
 
 import (
 	"context"
 	"fmt"
-	"os"
-	"os/signal"
-	"syscall"
+	"io"
 
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/pkg/rootless"
 	"github.com/moby/buildkit/util/apicaps"
-	"github.com/moby/sys/reexec"
-	"github.com/moby/term"
 	"github.com/spf13/cobra"
 )
 
 var honorXDG bool
 
-func newDaemonCommand() (*cobra.Command, error) {
+func newDaemonCommand(stderr io.Writer) (*cobra.Command, error) {
 	// FIXME(thaJeztah): config.New also looks up default binary-path, but this code is also executed when running "--version".
 	cfg, err := config.New()
 	if err != nil {
@@ -42,7 +38,7 @@ func newDaemonCommand() (*cobra.Command, error) {
 			}
 			if opts.Validate {
 				// If config wasn't OK we wouldn't have made it this far.
-				_, _ = fmt.Fprintln(os.Stderr, "configuration OK")
+				_, _ = fmt.Fprintln(stderr, "configuration OK")
 				return nil
 			}
 
@@ -100,39 +96,36 @@ func init() {
 	honorXDG = rootless.RunningWithRootlessKit()
 }
 
-func main() {
-	if reexec.Init() {
-		return
-	}
-	ctx := context.Background()
+// Runner is used to run the daemon command
+type Runner interface {
+	Run(context.Context) error
+}
 
-	// Ignore SIGPIPE events. These are generated by systemd when journald is restarted while
-	// the docker daemon is not restarted and also running under systemd.
-	// Fixes https://github.com/docker/docker/issues/19728
-	signal.Ignore(syscall.SIGPIPE)
+type daemonRunner struct {
+	*cobra.Command
+}
 
-	// Set terminal emulation based on platform as required.
-	_, stdout, stderr := term.StdStreams()
-	onError := func(err error) {
-		fmt.Fprintf(stderr, "%s\n", err)
-		os.Exit(1)
-	}
+func (d daemonRunner) Run(ctx context.Context) error {
+	configureGRPCLog(ctx)
 
-	// initial log formatting; this setting is updated after the daemon configuration is loaded.
+	return d.ExecuteContext(ctx)
+}
+
+// NewDaemonRunner creates a new daemon runner with the given
+// stdout and stderr writers.
+func NewDaemonRunner(stdout, stderr io.Writer) (Runner, error) {
 	err := log.SetFormat(log.TextFormat)
 	if err != nil {
-		onError(err)
+		return nil, err
 	}
 
 	initLogging(stdout, stderr)
-	configureGRPCLog()
 
-	cmd, err := newDaemonCommand()
+	cmd, err := newDaemonCommand(stderr)
 	if err != nil {
-		onError(err)
+		return nil, err
 	}
 	cmd.SetOut(stdout)
-	if err := cmd.ExecuteContext(ctx); err != nil {
-		onError(err)
-	}
+
+	return daemonRunner{cmd}, nil
 }
