+++ daemon/command/main_linux_test.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"testing"
